requirements.txt
langchain_community
streamlit
python-dotenv
langchain_core
pypdf
langchain_classic
langchain_groq
faiss-cpu